backend/requirement.txt
fastapi
uvicorn
transformers
sentence-transformers
chromadb
python-dotenv
newspaper3k
beautifulsoup4
playwright
